Moves version sync function to setup.py.
Moves the version sync function to setup.py to
resolve an import error for IPython during build
time.
